+++ frontend/src/components/Hazard/SeismicHazard/HazardCurveMetadata.js
@@ -3,23 +3,25 @@ import React, { useState, useEffect, useContext } from "react";
 import { GlobalContext } from "context";
 
 const HazardCurveMetadata = ({ selectedEnsemble, selectedIM, vs30 }) => {
-  const { selectedZFactor, selectedSoilClass, isNZCodeComputed } = useContext(
-    GlobalContext
-  );
+  const {
+    selectedZFactor,
+    selectedSoilClass,
+    isNZS1170p5Computed,
+  } = useContext(GlobalContext);
 
   const [localZFactor, setLocalZFactor] = useState(1);
   const [localSoilClass, setLocalSoilClass] = useState(1);
 
   /* 
-    We only update Z Factor and Soil Class after user click Compute as then we get updated NZ Code based on those values.
+    We only update Z Factor and Soil Class after user click Compute as then we get updated NZS1170P5 code based on those values.
     Instead of updating everytime users change them
   */
   useEffect(() => {
-    if (isNZCodeComputed === true) {
+    if (isNZS1170p5Computed === true) {
       setLocalZFactor(selectedZFactor);
       setLocalSoilClass(selectedSoilClass);
     }
-  }, [isNZCodeComputed]);
+  }, [isNZS1170p5Computed]);
 
   return (
     <div className="form-group">