+++ frontend/src/components/PermissionConfig/EditUserPermission.js
@@ -84,7 +84,8 @@ const EditUserPermission = () => {
         setUserDataFetching(true);
 
         await fetch(
-          CONSTANTS.CORE_API_BASE_URL + CONSTANTS.INTERMEDIATE_API_AUTH0_USERS_ENDPOINT,
+          CONSTANTS.CORE_API_BASE_URL +
+            CONSTANTS.INTERMEDIATE_API_AUTH0_USERS_ENDPOINT,
           {
             headers: {
               Authorization: `Bearer ${token}`,